% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-doc.r
\docType{package}
\name{bittrex-package}
\alias{bittrex-package}
\alias{bittrex}
\title{Client for the Bittrex Crypto-Currency Exchange}
\description{
This software is in no way affiliated, endorsed, or approved by
the Bittrex crypto-currency exchange or any of its affiliates. It comes with
absolutely no warranty and should not be used in actual trading
unless the user can read and understand the source and knows what they are
doing.

The \code{bittrex} package is an R implementation of the REST interface used
by the Bittrex crypto-currency exchange (\url{https://bittrex.com}). It
provides functions for all endpoints currently (as of May 16, 2017)
supported by the exchange. This includes the ability to retrieve price,
volume, and orderbook information as well as the ability to trade
crypto-currencies.

Calls to the exchange are categorized as either public, which includes
requests for price, volume, and order book information, and private, which
includes all requests requiring an account including placing buy or sell
orders. Public calls can be used immediately after installing the package.
Private calls require creating an account at \url{https://bittrex.com} and
creating API and secret keys with appropriate permissions.

Private calls retrieve the API and secret key using the BITTREX_API_KEY and
BITTREX_SECRET_KEY environment variables. These may be set by the user
before opening the R session or, they can be set using the
\code{\link[=bt_authenticate]{bt_authenticate()}} function.
}
\section{Public Function Calls}{

\itemize{
\item{\code{\link[=bt_api_check]{bt_api_check()}}  }{check if the bittrex REST API is working}
\item{\code{\link[=bt_getcurrencies]{bt_getcurrencies()}}  }{all supported currencies at Bittrex along with other meta data}
\item{\code{\link[=bt_getmarkethistory]{bt_getmarkethistory()}}  }{the latest trades that have occurred for a specified market}
\item{\code{\link[=bt_getmarkets]{bt_getmarkets()}}  }{the open and available trading markets at Bittrex along with other meta data}
\item{\code{\link[=bt_getmarketsummary]{bt_getmarketsummary()}}  }{the last 24 hours' summary of a specific market}
\item{\code{\link[=bt_getmarketsummaries]{bt_getmarketsummaries()}}  }{the last 24 hours' summary of all active markets}
\item{\code{\link[=bt_getorderbook]{bt_getorderbook()}}  }{the orderbook for a given market}
\item{\code{\link[=bt_getticker]{bt_getticker()}}  }{the current tick values for a market}
}
}

\section{Private Function Calls}{

\itemize{
\item{\code{\link[=bt_authenticate]{bt_authenticate()}}  }{provide user authentication data}
\item{\code{\link[=bt_buy]{bt_buy()}}  }{place a buy limit order}
\item{\code{\link[=bt_cancel]{bt_cancel()}}  }{cancel a buy or sell order}
\item{\code{\link[=bt_getbalances]{bt_getbalances()}}  }{account balances for all currencies}
\item{\code{\link[=bt_getbalance]{bt_getbalance()}}  }{account balance for a specified currency}
\item{\code{\link[=bt_getdepositaddress]{bt_getdepositaddress()}}  }{retrieve or generate an address for a specified
currency}
\item{\code{\link[=bt_getdeposithistory]{bt_getdeposithistory()}}  }{retrieve your deposit history}
\item{\code{\link[=bt_getopenorders]{bt_getopenorders()}}  }{retrieve data for all open orders}
\item{\code{\link[=bt_getorder]{bt_getorder()}}  }{retrieve a single order by uuid}
\item{\code{\link[=bt_getorderhistory]{bt_getorderhistory()}}  }{recent order history for an account }
\item{\code{\link[=bt_getwithdrawalhistory]{bt_getwithdrawalhistory()}}  }{retrieve your withdrawal history}
\item{\code{\link[=bt_sell]{bt_sell()}}  }{place a sell limit order}
\item{\code{\link[=bt_withdraw]{bt_withdraw()}}  }{withdraw funds from your account}
}
}

\references{
\url{https://bittrex.com},
\url{https://github.com/kaneplusplus/bittrex}
}
