% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bittrex.r
\name{bittrex_api_check}
\alias{bittrex_api_check}
\title{Check the Connection to the Bittrex Exchange}
\usage{
bittrex_api_check(warn = TRUE)
}
\arguments{
\item{warn}{if the request is not successful, should a warning be provided
with the status code.}
}
\value{
A named logical indicating if you can connect to the exchange
through the public interface.
}
\description{
The \code{bittrex_api_check} function checks to see
if you can sucessfully connect to the 
Bittrex crypto-currency exchange 
\url{https://bittrex.com}.
}
\examples{
\dontrun{
bittrex_api_check()
}
}
