% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bittrex.r
\name{bittrex_authenticate}
\alias{bittrex_authenticate}
\title{Provide User Authentication Data}
\usage{
bittrex_authenticate(api_key, secret_key)
}
\arguments{
\item{api_key}{the api key provided by the exchange}

\item{secret_key}{the secret key provided by the exchange}
}
\description{
The \code{bittrex_authenicate} function sets the 
BITTREX_API_KEY and BITTREX_SECRET_KEY environment variables in your current
session to access your account information on the Bittrex crypto-currency
exchange (\url{https://bittrex.com}).
}
