% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private-interface.r
\name{bt_getbalance}
\alias{bt_getbalance}
\title{Account Balance for a Specified Currency}
\usage{
bt_getbalance(currency)
}
\arguments{
\item{currency}{currency to retrieve the account balance for.}
}
\value{
A named list with the following elements:
\itemize{
\item{success: }{a boolean indicating whether the request was successful.}
\item{message: }{a string describing the error if the request was not
successful, otherwise an empty string.}
\item{result:  }{a \code{data.frame} with the currency, balance,
available funds, the amount of any pending transactions, and
cryptographic addresses that can be used to receive funding.
}
}
}
\description{
The \code{bt_getbalance()} function retrieves the account balance
for a specified currency on \url{https://bittrex.com}. This function
can only be used after you provide information for authentication.
}
\examples{
\dontrun{
Note you must authenticate first.
bt_getbalance("btc")$result
#   currency balance available pending                     crypto_address
# 1      BTC       0         0       0 1Q6WissSMNF7NCNw3sDXQ2F7AbrSCYouj2
}
}
\references{
https://bittrex.com/api/v1.1/account/getbalances
}
\seealso{
\code{\link[=bt_authenticate]{bt_authenticate()}}
}
