% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private-interface.r
\name{bt_getdeposithistory}
\alias{bt_getdeposithistory}
\title{Retrieve Deposit History}
\usage{
bt_getdeposithistory(currency)
}
\arguments{
\item{currency}{(optional) the currency to retrieve the deposits for. If this
is not specified then deposit history for all currencies is retrieved.}
}
\value{
A named list with the following elements:
\itemize{
 \item{success: }{a boolean indicating if the request successful?}
 \item{message: }{a string describing the error if the request was not 
                  successful, otherwise and empty string."}
 \item{result:  }{a \code{data.frame} providing data about 
   previously completed orders including the order uuid, the currency,
   the time of the withdraw, the quantity, etc.
 }
}
}
\description{
The \code{bt_getdeposithistory} function retrieves the
deposit history for an account on the Bittrex crypto-currency exchange
\url{https://bittrex.com}. This function can be used after you 
provide authentication information.
}
\examples{
\dontrun{
bt_getdeposithistory()
}
}
\references{
\url{https://bittrex.com/api/v1.1/account/getdeposithistory?apikey=API_KEY?currency=BTC}
}
\seealso{
\code{link{bt_authenticate}}
}
