% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bittrex.r
\name{buy}
\alias{buy}
\title{Place a Buy Limit Order}
\usage{
buy(market, quantity, type = c("market", "limit"), rate)
}
\arguments{
\item{market}{the market to place the buy limit order on.}

\item{quantity}{the quantity of the transaction currency to buy.}

\item{rate}{the price you are willing to pay per unit of the 
transaction currency.}
}
\value{
A named list with the following elements:
\itemize{
 \item{success: }{a boolean indicating if the request successful?}
 \item{message: }{a string describing the error if the request was not 
                  successful, otherwise and empty string."}
 \item{result:  }{a named list, called "uuid" whose element is an integer
   identifying the order. This value is used to query the status of
   of the order with either the \code{link{getorder}} or 
   \code{link{getopenorders}} function. When the order is fulfilled it
   appears in the order history \code{data.frame} returned by the
   \code{link{getorderhistory}} function.
 }
}
}
\description{
The \code{buy} function places a buy order onto the 
Bittrex crypto-currency exchange \url{https://bittrex.com}. This function
only works after you have set up authentication.
}
\examples{
\dontrun{
# Buy one litecoin for 1169 bitcoins. 
order = buy("btc-ltc", 1, 0.0001)
}
}
\references{
\url{https://bittrex.com/api/v1.1/account/sell}
}
\seealso{
\code{link{bittrex_authenticate}} \code{\link{sell}}
\code{\link{getorder}} \code{\link{getopenorders}} 
\code{\link{getorderhistory}}
}
