% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bittrex.r
\name{getbalance}
\alias{getbalance}
\title{Account Balance for a Specified Currency}
\usage{
getbalance(currency)
}
\arguments{
\item{currency}{currency to retrieve the account balance for.}
}
\value{
A named list with the following elements:
\itemize{
 \item{success: }{a boolean indicating if the request successful?}
 \item{message: }{a string describing the error if the request was not 
                  successful, otherwise and empty string.}
 \item{result:  }{a \code{data.frame} with the currency, balance, 
   available funds, the amount of any pending transactions, and 
   crypographic addresses that can be used to receive funding.
 }
}
}
\description{
The \code{getbalance} function retrieves the account balance
for a specified currency on the Bittrex crypto-currency 
exchange \url{https://bittrex.com}. This function
can be used after you provide information for authentication.
}
\examples{
\dontrun{
balances = getbalance("btc")$result
}
}
\references{
\url{https://bittrex.com/api/v1.1/account/getbalances}
}
\seealso{
\code{\link{bittrex_authenticate}}
}
