% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bittrex.r
\name{getdepositaddress}
\alias{getdepositaddress}
\title{Retrieve the Address for a Specified Currency}
\usage{
getdepositaddress(currency)
}
\arguments{
\item{currency}{currency to get the deposit address}
}
\value{
A named list with the following elements:
\itemize{
 \item{success: }{a boolean indicating if the request successful?}
 \item{message: }{a string describing the error if the request was not 
                  successful, otherwise and empty string. If an 
                  address has not been generated this field will have value
                  "ADDRESS_GENERATING" until it is available.}
 \item{result:  }{a \code{data.frame} with the one row and columns 
                  providing the currency and the address.}
}
}
\description{
The \code{getdepositaddress} retrieves or creates the account 
deposit address for a specified currency.
}
\examples{
\dontrun{
# Get the address of the bitcoin currency.
getdepositaddress("btc")
}
}
\references{
\url{https://bittrex.com/api/v1.1/account/getdepositaddress?apikey=API_KEY&currency=VTC}
}
