% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bittrex.r
\name{getopenorders}
\alias{getopenorders}
\title{Order Data for all Open Orders}
\usage{
getopenorders(market)
}
\arguments{
\item{market}{(optional) the market on which you would like to see all 
open orders. If not specified, then all open orders
for all markets are returned.}
}
\value{
A named list with the following elements:
\itemize{
 \item{success: }{a boolean indicating if the request successful?}
 \item{message: }{a string describing the error if the request was not 
                  successful, otherwise and empty string."}
 \item{result:  }{a \code{data.frame} providing information about the 
   open orders including (but not limited to) the market, quantity remaining
   in the order, the type of order, and when the order was opened.
 }
}
}
\description{
The \code{getopenorders} function retrieves all open orders
on the Bittrex crypto-currency 
exchange \url{https://bittrex.com}. This function
can be used after you provide information for authentication.
}
\examples{
\dontrun{
getopenorders()
}
}
\references{
\url{https://bittrex.com/api/v1.1/market/getopenorders}
}
\seealso{
\code{\link{bittrex_authenticate}}
}
