% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bittrex.r
\name{getticker}
\alias{getticker}
\title{Get the Ticker Values for a Market}
\usage{
getticker(market)
}
\value{
A named list with the following elements:
\itemize{
 \item{success: }{a boolean indicating if the request successful?}
 \item{message: }{a string describing the error if the request was not 
                  successful, otherwise and empty string.}
 \item{result:  }{A \code{data.frame} with the bid, ask, and last 
                  trasaction price.}
}
}
\description{
The \code{getticker} function returns the bid, ask, and last
transaction price for a specified market on the Bittrex crypto-currency 
exchange (\url{https://bittrex.com}).
}
\examples{
\dontrun{
getticker("btc-ltc")
head(markets)
}
}
\references{
\url{https://bittrex.com/api/v1.1/public/getticker}
}
