% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bittrex.r
\name{getwithdrawalhistory}
\alias{getwithdrawalhistory}
\title{Retrieve Withdrawal History}
\usage{
getwithdrawalhistory(currency)
}
\arguments{
\item{currency}{(optional) the currency to retrieve the withdraw for. If this
is not specified then withdraw history for all currencies is retrieved.}
}
\value{
A named list with the following elements:
\itemize{
 \item{success: }{a boolean indicating if the request successful?}
 \item{message: }{a string describing the error if the request was not 
                  successful, otherwise and empty string."}
 \item{result:  }{a \code{data.frame} providing data about 
   previously completed orders including the order uuid, the currency,
   the time of the withdraw, the quantity, etc.
 }
}
}
\description{
The \code{getwithdrawalhistory} function retrieves the
withdraw history for an account on the Bittrex crypto-currency exchange
\url{https://bittrex.com}. This function can be used after you 
provide authentication information.
}
\examples{
\dontrun{
getwithdrawalhistory()
}
}
\references{
\url{https://bittrex.com/api/v1.1/account/getwithdrawalhistory?apikey=API_KEY?currency=BTC}
}
\seealso{
\code{link{bittrex_authenticate}}
}
