% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bittrex.r
\name{sell}
\alias{sell}
\title{Place a Sell Limit Order}
\usage{
sell(market, quantity, rate, type = c("market", "limit"))
}
\arguments{
\item{market}{the market to place the buy limit order on.}

\item{quantity}{the quantity of the reference currency to sell.}

\item{rate}{the price you would like to get per unit of the 
transaction 
currency.}

\item{type}{either "market" or "limit". Note that market orders are currently
disabled.}
}
\value{
A named list with the following elements:
\itemize{
 \item{success: }{a boolean indicating if the request successful?}
 \item{message: }{a string describing the error if the request was not 
                  successful, otherwise and empty string."}
 \item{result:  }{a named list, called "uuid" whose element is an integer
   identifying the order. This value is used to query the status of
   of the order with either the \code{link{getorder}} or 
   \code{link{getopenorders}} function. When the order is fulfilled it
   appears in the order history \code{data.frame} returned by the
   \code{link{getorderhistory}} function.
 }
}
}
\description{
The \code{sell} function places a buy order onto the 
C-Cex crypto-currency exchange \url{https://bittrex.com}. This function
only works if you have set up authentication. 

NOTE: market orders are currently disabled.
}
\examples{
\dontrun{
# Sell one litecoin for 0.000001 bitcoins. 
order = selllimit("btc-ltc", 1, 0.000001)
}
}
\references{
\url{https://bittrex.com/api/v1.1/market/selllimit}
}
\seealso{
\code{link{bittrex_authenticate}} \code{\link{buy}}
\code{\link{getopenorders}} \code{\link{getorderhistory}}
}
