% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private-interface.r
\name{bt_cancel}
\alias{bt_cancel}
\title{Cancel an Open Order}
\usage{
bt_cancel(uuid)
}
\arguments{
\item{uuid}{the uuid of the order you would like to cancel.}
}
\value{
A named list with the following elements:
\itemize{
 \item{success: }{a boolean indicating if the request successful?}
 \item{message: }{a string describing the error if the request was not 
                  successful, otherwise and empty string.}
 \item{result:  }{always NULL}
}
}
\description{
The \code{bt_cancel} function cancels an open order on the
C-Cex crypto-currency exchange \url{https://bittrex.com}. This function
is called after providing information to authenticate your account and 
after an order is placed using either 
the \code{link{bt_buy}} or \code{link{bt_sell}} functions.
}
\examples{
\dontrun{
bt_cancel(uuid) 
}
}
\references{
\url{https://bittrex.com/api/v1.1/account/cancel}
}
\seealso{
\code{\link{bt_authenticate}} \code{\link{bt_getopenorders}}
}
