% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private-interface.r
\name{bt_getopenorders}
\alias{bt_getopenorders}
\title{Order Data for all Open Orders}
\usage{
bt_getopenorders(market)
}
\arguments{
\item{market}{(optional) the market on which you would like to see all
open orders. If not specified, then all open orders
for all markets are returned.}
}
\value{
A named list with the following elements:
\itemize{
\item{success: }{a boolean indicating whether the request was successful.}
\item{message: }{a string describing the error if the request was not
successful, otherwise an empty string.}
\item{result:  }{a \code{data.frame} providing information about the
open orders including (but not limited to) the market, quantity remaining
in the order, the type of order, and when the order was opened.
}
}
}
\description{
The \code{bt_getopenorders()} function retrieves all the user's open
orders on \url{https://bittrex.com}. This function can only be used after
you provide information for authentication.
}
\examples{
\dontrun{
# Note you must authenticate first.
bt_getopenorders()
# $success
# [1] TRUE
#
# $message
# [1] ""
#
# $result
#   uuid                           order_uuid exchange order_type quantity
# 1   NA 2d6169e9-17fb-4f2a-8aff-37418b515624  ETH-LTC  LIMIT_BUY        1
#   quantity_remaining limit commission_paid price price_per_unit
# 1                  1 0.001               0     0             NA
#                   opened closed cancel_initiated immediate_or_cancel
# 1 2017-07-11T18:53:30.07     NA            FALSE               FALSE
#   is_conditional condition condition_target
# 1          FALSE      NONE               NA
}
}
\references{
https://bittrex.com/api/v1.1/market/getopenorders
}
\seealso{
\code{\link[=bt_authenticate]{bt_authenticate()}}
}
