% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private-interface.r
\name{bt_getorder}
\alias{bt_getorder}
\title{Order Data for a Specified Order}
\usage{
bt_getorder(uuid)
}
\arguments{
\item{uuid}{the uuid of the order.}
}
\value{
A named list with the following elements:
\itemize{
 \item{success: }{a boolean indicating if the request successful?}
 \item{message: }{a string describing the error if the request was not 
                  successful, otherwise and empty string."}
 \item{result:  }{a \code{data.frame} providing information about the 
   open order including (but not limited to) the market, quantity remaining
   in the order, the type of order, and when the order was opened.
 }
}
}
\description{
The \code{bt_getorder} function retrieves open order data 
on the Bittrex crypto-currency 
exchange \url{https://bittrex}. This function
can be used after you provide information for authentication.
}
\examples{
\dontrun{
# Note you must authenticate and define a uuid first.
bt_getorder(uuid)
# $success
# [1] TRUE
# 
# $message
# [1] ""
# 
# $result
#   account_id                           order_uuid exchange      type quantity
# 1         NA 63181c27-dd14-476e-960c-1bd8366bb312  ETH-LTC LIMIT_BUY        1
#   quantity_remaining limit reserved reserve_remaining commission_reserved
# 1                  1 0.001    0.001             0.001             2.5e-06
#   commission_reserve_remaining commission_paid price price_per_unit
# 1                      2.5e-06               0     0             NA
#                    opened closed is_open                             sentinel
# 1 2017-07-11T17:05:55.583     NA    TRUE 78b210fa-a156-4f4d-8043-51163740056f
#   cancel_initiated immediate_or_cancel is_conditional condition
# 1            FALSE               FALSE          FALSE      NONE
#   condition_target
# 1               NA
}
}
\references{
\url{https://bittrex.com/api/v1.1/account/getorder&apikey=API_KEY&uuid=0cb4c4e4-bdc7-4e13-8c13-430e587d2cc1}
}
\seealso{
\code{\link{bt_authenticate}} \code{\link{bt_getopenorders}}.
}
