% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public-interface.r
\name{bt_getticker}
\alias{bt_getticker}
\title{Get the Ticker Values for a Market}
\usage{
bt_getticker(market)
}
\arguments{
\item{market}{the market to get the ticker for.}
}
\value{
A named list with the following elements:
\itemize{
 \item{success: }{a boolean indicating if the request successful?}
 \item{message: }{a string describing the error if the request was not 
                  successful, otherwise and empty string.}
 \item{result:  }{A \code{data.frame} with the bid, ask, and last 
                  trasaction price.}
}
}
\description{
The \code{bt_getticker} function returns the bid, ask, and last
transaction price for a specified market on the Bittrex crypto-currency 
exchange (\url{https://bittrex.com}). The complete list of 
markets is available via the \code{\link{bt_getmarkets}} function.
}
\examples{
\dontrun{
getticker("btc-ltc")
head(markets)
}
}
\references{
\url{https://bittrex.com/api/v1.1/public/getticker}
}
\seealso{
\code{\link{bt_getmarkets}}
}
